fix(feedback) restructure component folder structure, fix all naming issues, fix all js docs issues, fix issue in readme
[feedback]
