+++ client/components/App.jsx
@@ -1,10 +1,24 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import NavigationBar from './General/NavigationBar.jsx';
-import FlashMessageList from './flashMessages/flashMessagesList.jsx';
-import Footer from './General/footer.jsx';
+import NavigationBar from './includes/NavigationBar.jsx';
+import FlashMessageList from './includes/FlashMessagesList.jsx';
+import Footer from './includes/Footer.jsx';
 
+/**
+ * The App page
+ *
+ * @class App
+ * @extends {React.Component}
+ */
 class App extends React.Component {
+
+  /**
+   * This method renders the component
+   *
+   * @returns {Object} jsx component
+   *
+   * @memberof App
+   */
   render() {
     return (
       <div className="teal lighten-5 wrapper">
