--- client/test/store/store.spec.js
@@ -1,4 +1,0 @@
-// import redux from 'react-redux';
-
-// const Dispatcher = redux.Dispatch;
-
