+++ package.json
@@ -13,8 +13,8 @@
     "lint:watch": "npm run lint -- --watch",
     "db:test:seed": "NODE_ENV=test babel-node ./server/test-utility/seeders/seedDatabase",
     "db:seed": "NODE_ENV=development babel-node ./server/test-utility/seeders/seedDatabase",
-    "mocha:testServer": "NODE_ENV=test nyc --reporter=html mocha \"./server/test/**/*.spec.js\" --compilers js:babel-register --timeout 10000",
-    "mocha:testClient": "nyc --reporter=html mocha client/test/testSetup.js \"./client/test/**/*.spec.js\"",
+    "mocha:testServer": "NODE_ENV=test nyc --reporter=text mocha \"./server/test/**/*.spec.js\" --compilers js:babel-register --timeout 5000",
+    "mocha:testClient": "nyc --reporter=text mocha client/test/testSetup.js \"./client/test/**/*.spec.js\"",
     "mocha:testClientWatch": "npm run mocha:testClient -- --watch",
     "db:migrate": "NODE_ENV=development sequelize db:migrate:undo:all && sequelize db:migrate",
     "test:server": "NODE_ENV=test npm run db:test:seed && npm run mocha:testServer",
