+++ server/app.js
@@ -1,32 +1,12 @@
 import express from 'express';
 import path from 'path';
-// import webpack from 'webpack';
-// import webpackHotMiddleware from 'webpack-hot-middleware';
-// import webpackMiddleware from 'webpack-dev-middleware';
 import bodyParser from 'body-parser';
 import Route from './routes';
-
-// let webpackConfig;
-// if (process.env.NODE_ENV === 'development') {
-//   webpackConfig = require('../webpack.config.dev');
-// } else {
-//   webpackConfig = require('../../webpack.config.prod');
-// }
 
 require('dotenv')
   .config();
 
 const app = express();
-// const compiler = webpack(webpackConfig);
-
-// app.use(webpackMiddleware(compiler, {
-//   hot: true,
-//   publicPath: webpackConfig.output.publicPath,
-//   noInfo: true,
-// }));
-
-// app.use(webpackHotMiddleware(compiler));
-// Log requests to the console.
 
 // Parse incoming requests data (https://github.com/expressjs/body-parser)
 app.use(bodyParser.json());
